requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.2
django-cache-url==3.4.5
environs==10.3.0
marshmallow==3.20.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
